move mapset + process_chain models
